+++ geocoding_android/lib/geocoding_android.dart
@@ -13,17 +13,24 @@ class GeocodingAndroid extends GeocodingPlatform {
   }
 
   @override
+  Future<void> setLocaleIdentifier(
+    String localeIdentifier,
+  ) {
+    final parameters = <String, String>{
+      'localeIdentifier': localeIdentifier,
+    };
+
+    return _channel.invokeMethod('setLocaleIdentifier', parameters);
+  }
+
+  @override
   Future<List<Location>> locationFromAddress(
-    String address, {
-    String? localeIdentifier,
-  }) async {
+    String address,
+  ) async {
     final parameters = <String, String>{
       'address': address,
     };
 
-    if (localeIdentifier != null) {
-      parameters['localeIdentifier'] = localeIdentifier;
-    }
     try {
       final placemarks = await _channel.invokeMethod(
         'locationFromAddress',
@@ -40,21 +47,37 @@ class GeocodingAndroid extends GeocodingPlatform {
   @override
   Future<List<Placemark>> placemarkFromCoordinates(
     double latitude,
-    double longitude, {
-    String? localeIdentifier,
-  }) async {
+    double longitude,
+  ) async {
     final parameters = <String, dynamic>{
       'latitude': latitude,
       'longitude': longitude,
     };
 
-    if (localeIdentifier != null) {
-      parameters['localeIdentifier'] = localeIdentifier;
-    }
-
     final placemarks =
         await _channel.invokeMethod('placemarkFromCoordinates', parameters);
     return Placemark.fromMaps(placemarks);
+  }
+
+  @override
+  Future<List<Placemark>> placemarkFromAddress(
+    String address,
+  ) async {
+    final parameters = <String, String>{
+      'address': address,
+    };
+
+    try {
+      final placemarks = await _channel.invokeMethod(
+        'placemarkFromAddress',
+        parameters,
+      );
+
+      return Placemark.fromMaps(placemarks);
+    } on PlatformException catch (e) {
+      _handlePlatformException(e);
+      rethrow;
+    }
   }
 
   void _handlePlatformException(PlatformException platformException) {